Fix compilation problems for Open CV 4.0.0-beta
